+++ primitives/src/lib.rs
@@ -21,7 +21,6 @@
 extern crate alloc;
 
 mod asset;
-pub mod checked_ops_res;
 pub mod constants;
 pub mod fixed;
 mod market;
