+++ primitives/src/math/mod.rs
@@ -15,6 +15,5 @@
 // You should have received a copy of the GNU General Public License
 // along with Zeitgeist. If not, see <https://www.gnu.org/licenses/>.
 
-pub mod check_arithm_rslt;
-mod consts;
+pub mod checked_ops_res;
 pub mod fixed;
